End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -53,12 +54,13 @@ public class AuctionsController : ControllerBase
         return Ok(_mapper.Map<AuctionDto>(auction));
     }
     
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: Add user to seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
         
         _context.Auctions.Add(auction);
         
@@ -77,6 +79,7 @@ public class AuctionsController : ControllerBase
             new { auction.Id }, newAuction);
     }
     
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -87,7 +90,10 @@ public class AuctionsController : ControllerBase
             return NotFound();
         }
         
-        // TODO: Validate that user is the original seller
+        if(auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
         
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -106,6 +112,7 @@ public class AuctionsController : ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -115,7 +122,10 @@ public class AuctionsController : ControllerBase
             return NotFound();
         }
 
-        // TODO: Validate that user is the original seller
+        if(auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
         _context.Auctions.Remove(auction);
         
         await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
